Remove old master files
