+++ client/annotation_client.py
@@ -14,7 +14,7 @@ import client.utils as utils
 from client.client_config import ClientConfig
 from client.screens.common import Alert
 from client.screens.image_view_screen import ImageViewScreen
-from client.screens.instance_annnotator_screen import InstanceAnnotatorScreen
+from client.screens.instance_annotator_screen import InstanceAnnotatorScreen
 from client.screens.project_select_screen import ProjectSelectScreen
 from client.screens.project_tool_screen import ProjectToolScreen
